Seperate constants from libs
