+++ archetypes/Portfolio/History/index.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from 'react';
 import { useRouter } from 'next/router';
 import { ethers } from 'ethers';
 import { CommitActionEnum } from '@tracer-protocol/pools-js';
-import { ARBITRUM, CommitActionToQueryFocusMap } from '@libs/constants';
+import { ARBITRUM, CommitActionToQueryFocusMap } from '@constants/index';
 import { useWeb3 } from '@context/Web3Context/Web3Context';
 import { Table, TableHeader, TableHeaderCell, TableRow, TableRowCell } from '@components/General/TWTable';
 import Pagination, { PageNumber } from '@components/General/Pagination';
