+++ archetypes/Portfolio/Queued/index.tsx
@@ -1,6 +1,6 @@
 import React, { useMemo, useState } from 'react';
 import { CommitEnum, CommitActionEnum, SideEnum } from '@tracer-protocol/pools-js';
-import { CommitActionToQueryFocusMap } from '@libs/constants';
+import { CommitActionToQueryFocusMap } from '@constants/commits';
 import { useWeb3 } from '@context/Web3Context/Web3Context';
 import { Table, TableHeader, TableHeaderCell, TableRow, TableRowCell } from '@components/General/TWTable';
 import TWButtonGroup from '@components/General/TWButtonGroup';
