+++ constants/general.ts
@@ -1,0 +1,3 @@
+import { ethers } from 'ethers';
+
+export const MAX_SOL_UINT = ethers.BigNumber.from('340282366920938463463374607431768211455');
