+++ constants/networks.ts
@@ -1,0 +1,6 @@
+// Networks
+export const ARBITRUM_RINKEBY = '421611';
+export const ARBITRUM = '42161';
+export const MAINNET = '1';
+export const RINKEBY = '4';
+export const KOVAN = '42';
